fix(ui): order backups servers list by rank like the group page
Extract the rank-bucket grouping from GroupDetail into a shared
groupServersByRank helper in types.ts, reuse it for the ServerDetail
sibling list (which had its own inline copy), and apply it to the
backups panel's Servers table with rank subheader rows.
